+++ hw01/src/main/java/ru/otus/spring/Main.java
@@ -1,8 +1,14 @@
 package ru.otus.spring;
 
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+import ru.otus.spring.dao.QuestionsDao;
+
 public class Main {
 
     public static void main(String[] args) {
-        QuestionModule.runTest();
+        ApplicationContext applicationContext = new ClassPathXmlApplicationContext("spring-context.xml");
+        var questionsDao = applicationContext.getBean(QuestionsDao.class);
+        questionsDao.runTest();
     }
 }
