+++ hw01/src/main/java/ru/otus/spring/dao/QuestionsDao.java
@@ -4,9 +4,10 @@ import java.util.List;
 import java.util.Map;
 
 @SuppressWarnings("checkstyle:LeftCurly")
-public interface QuestionsDao
-{
+public interface QuestionsDao {
     Map<String,List<String>> getAll();
 
     List<String> findByQuestions(String questionName);
+
+    void runTest();
 }
