+++ hw01/src/main/resources/spring-context.xml
@@ -4,8 +4,12 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans
         http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="questionsDao" class="ru.otus.spring.dao.QuestionsDaoImpl">
+    <bean id="fileManager" class="ru.otus.spring.utils.FileManager">
         <constructor-arg index="0" value="questions.csv"/>
+    </bean>
+
+    <bean id="questionsDao" class="ru.otus.spring.dao.QuestionsDaoImpl">
+        <constructor-arg index="0" ref="fileManager"/>
     </bean>
 
 </beans>
