+++ hw01/src/test/java/QuestionDaoTest.java
@@ -1,6 +1,7 @@
 import org.junit.jupiter.api.*;
 import ru.otus.spring.dao.QuestionsDao;
 import ru.otus.spring.dao.QuestionsDaoImpl;
+import ru.otus.spring.utils.FileManager;
 
 @DisplayName("Тестирование класса-хранилища вопросов")
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
@@ -8,7 +9,7 @@ public class QuestionDaoTest {
     private QuestionsDao questionsDao;
     @BeforeAll
     void beforeAll(){
-        this.questionsDao = new QuestionsDaoImpl("questions_test.csv");
+        this.questionsDao = new QuestionsDaoImpl(new FileManager("questions_test.csv"));
     }
 
     @Test
